Refactor asset handling; remove unused AssetType class and update references in QueryBuilder and Asset classes
